--- a/ProtoPromise_Unity/Assets/Plugins/ProtoPromise/Core/Promises/Internal/DeferredInternal.cs
+++ b/ProtoPromise_Unity/Assets/Plugins/ProtoPromise/Core/Promises/Internal/DeferredInternal.cs
@@ -120,7 +120,6 @@ namespace Proto.Promises
                 protected override void MaybeDispose()
                 {
                     Dispose();
-                    WaitWhileProgressReporting();
                     ObjectPool.MaybeRepool(this);
                 }
 
@@ -189,7 +188,6 @@ namespace Proto.Promises
                 {
                     Dispose();
                     _cancelationRegistration = default(CancelationRegistration);
-                    WaitWhileProgressReporting();
                     ObjectPool.MaybeRepool(this);
                 }