Removed extra WaitWhileProgressReporting in deferred dispose.

Master:

|         Type | Pending |     Mean |     Error |    StdDev | Allocated | Survived |
|------------- |-------- |---------:|----------:|----------:|----------:|---------:|
|   AsyncAwait |    True | 2.421 μs | 0.0234 μs | 0.0219 μs |         - |    688 B |
| ContinueWith |    True | 2.738 μs | 0.0145 μs | 0.0136 μs |         - |    416 B |

This PR:

|         Type | Pending |     Mean |     Error |    StdDev | Allocated | Survived |
|------------- |-------- |---------:|----------:|----------:|----------:|---------:|
|   AsyncAwait |    True | 2.383 μs | 0.0185 μs | 0.0173 μs |         - |    688 B |
| ContinueWith |    True | 2.708 μs | 0.0129 μs | 0.0121 μs |         - |    416 B |